Reorganize provider
