+++ apps/web/src/auth/providers/google.ts
@@ -1,0 +1,11 @@
+import BaseGoogleProvider from "next-auth/providers/google";
+
+export function GoogleProvider() {
+  if (process.env.GOOGLE_CLIENT_ID && process.env.GOOGLE_CLIENT_SECRET) {
+    return BaseGoogleProvider({
+      clientId: process.env.GOOGLE_CLIENT_ID,
+      clientSecret: process.env.GOOGLE_CLIENT_SECRET,
+      allowDangerousEmailAccountLinking: true,
+    });
+  }
+}
